+++ tom_targets/filters.py
@@ -2,7 +2,8 @@ from django.conf import settings
 from django.db.models import Q
 import django_filters
 
-from tom_targets.models import Target, TargetList, TargetMatchManager
+from tom_targets.models import Target, TargetList
+from tom_targets.base_models import TargetMatchManager
 from tom_targets.utils import cone_search_filter
